+++ app/views/community/groups/index.html.slim
@@ -36,7 +36,7 @@
                 .inviteGroup
                     = link_to "ユーザを招待", invite_users_community_group_url(group), class: "btn btn-primary rounded-pill", remote: true
                 .inviteModal
-                = render "inviteModal"
+                / = render "inviteModal"
                 .dropdown
                     #dropdownMenuButton1 aria-expanded="false" data-bs-toggle="dropdown" type="button" 
                         i.fas.fa-ellipsis-v
