+++ app/views/community/index.html.slim
@@ -10,7 +10,7 @@
         .card-body
             h5.card-title グループ
             p.card-text 友人などと一緒に目標を立てたり、知らない人同士お互いに高めあったりできます！
-            a.btn.btn-primary href="./groups"  Go to groups
+            a.btn.btn-primary href="./community/groups"  Go to groups
 
     .card style=("width: 22rem;") 
         svg.bd-placeholder-img.card-img-top aria-label=("Placeholder: Image cap") focusable="false" height="180" preserveaspectratio=("xMidYMid slice") role="img" width="100%" xmlns="http://www.w3.org/2000/svg"
