+++ app/views/layouts/application.html.slim
@@ -20,7 +20,7 @@ html
             li.first
               | こんにちは。
               strong
-                a title="#{current_user.name}" href="/admin/users/#{current_user.id}" 
+                a title="#{current_user.name}" href="/users/#{current_user.id}" 
                   = current_user.name
                 | さん
             li
