+++ app/views/users/edit.html.slim
@@ -1,0 +1,21 @@
+- content_for :css do
+    = stylesheet_link_tag "users"
+
+.block
+    h2 あなたの写真
+    .profile-edit
+        .profile-image 
+            - if @user.image?
+                img.img-thumbnail.img-fluid src="#{@user.image.url}" class="image-thumbnail"
+            - else
+                .no-image-explanation 
+                    | アップロードされた画像はありません
+        .image-upload-area 
+            p 新しい画像を追加
+            | プロフィールに写真を追加するにはまず、コンピューターから写真を選択し、「写真をアップロード」をクリックします。
+            .mb-3
+                = form_with model: @user do |f|
+                    .image-select-field
+                        = f.file_field :image
+                    .submit-image
+                        = f.submit "写真をアップロード", class: "btn btn-primary"
