+++ db/migrate/20220125120330_add_image_to_users.rb
@@ -1,0 +1,5 @@
+class AddImageToUsers < ActiveRecord::Migration[5.2]
+  def change
+    add_column :users, :image, :string
+  end
+end
